+++ Mods/SOD.StockMarket/Implementation/Trade/TradeOrder.cs
@@ -1,0 +1,31 @@
+﻿namespace SOD.StockMarket.Implementation.Trade
+{
+    internal struct TradeOrder
+    {
+        internal int StockId { get; }
+        internal decimal Price { get; }
+        internal int Amount { get; }
+        internal OrderType OrderType { get; }
+        internal bool Completed { get; private set; }
+
+        internal TradeOrder(OrderType orderType, int stockId, decimal price, int amount)
+        {
+            StockId = stockId;
+            Price = price;
+            Amount = amount;
+            OrderType = orderType;
+            Completed = false;
+        }
+
+        internal void Complete()
+        {
+            Completed = true;
+        }
+    }
+
+    internal enum OrderType
+    {
+        Buy,
+        Sell
+    }
+}